Refactor LS enable
Moved code from org.eclipse.cdt.lsp.editor.editor.ui to
org.eclipse.cdt.lsp plug-in.
Reasons:
- The logic for the check whether to open a C/C++ source file in the LSP
based editor or not should be done in the org.eclipse.cdt.lsp plug-in,
since the org.eclipse.ui.ide.editorAssociationOverride extension point
is here as well.

- The org.eclipse.cdt.lsp.editor.editor.ui has to check whether the
prefer LSB Editor has been set in the project properties or not on a
given project.

- Replaced argument type Object in ICLanguageServerProvider.isEnabledFor
by IProject. This makes it clearly to implement a test.