Fix broken fallback flags
The fallback flags determination was broken:
- CMakeBuildConfiguration.getScannerInformation(IResource) needs a File
as input since the infos are stored in a map with a File as key. This
will fail if IResource is a Folder.
- ClangdFallbackManager.flags(IResource): IResource must be the initial
file, not the initialProject, because the
getScannerInformation(IResource) needs a file which can be found in the
compile_commands.json in the project. Otherwise the compiler settings
cannot be determined my the cmake bundle in CDT.