[#139] fix slow content provider
the symbols has not been cached and due to the disconnect of the file
from buffer every call of refreshTreeContentFromLS triggers a new AST
build on server side for the file. This can take a while for a large
codebase. The solution is to cache the symbols for a file in the
SymbolsContainer. The symbols will only be refreshed when the buffer is
dirty.

Replaced the deprecated DocumentSymbolWithFile by DocumentSymbolWithURI

fixes #139