[#150] Restart clangd when clangd settings changed
Ask the user to restart clangd when a clangd option has been modified in
the prefs or project properties. We stop the clangd language server.
When the user opens a new C/C++ file or an already opened C/C++ file in
the LSP based C/C++ editor gets into focus, a new LS will be started by
LSP4E.

Added a 'clangd options' group to the preference page. If any element in
this group will be modified, a info dialog pops up, when the 'Apply' or
'Apply and Close' button will be pressed. The dialog asks the user to
restart clangd.

fixes #150