Update to Tycho 4.0.6
A newer Tycho is needed to enable build and test against more
recent platform.

tycho-packaging-plugin is needed explicitly to ensure that
the headers in MANIFEST.MF are not removed during build. This
ensures cdt-lsp doesn't have this same problem as CDT had recently
https://github.com/eclipse-cdt/cdt/issues/551

The inclusion of tycho-packaging-plugin also contains the missing
source references in the MANIFEST.MF outputs.