Account for non-API method signature change in LSP4E
The signature of the LanguageServerWrapper restart method
changed between the last two snapshots. This means we
needed to remove our catch a few commits ago. This also
means we need to use a version of LSP4E >= the version
where the non-API changed, hence why we care about patch
version in the MANIFEST.MF.

We probably need to have LS4E expose needed methods as API
so we don't have problems mixing and matching versions going
forward.

See the corresponding LSP4E commit:

https://github.com/eclipse/lsp4e/commit/a27b7a93ad74b3edc95c11d4eaf29342a04489d9

And the corresponding CDT-LSP commit that fixed the compilation issue:

https://github.com/eclipse-cdt/cdt-lsp/commit/6d6209bfeeeed817c6c4a321f564e0f377552b3d