[#408] add clangd command line options validator
if clangd version >= 11, then the clangd binary will be called with
configured command line options plus --check and --log=error.
If the stderr stream remains empty everything is fine and we can go on.
Otherwise we through an IOException in the start() method. This leads to
an error dialog with the strings returned from stderr stream.

fixes #408