Implement "Try new editing experience" to promote CDT LSP
Contribute a banner to the CEditor and CLspEditor to promote
the new C/C++ editing experience based on LSP. The banner in
CEditor enables a quick switch to the LSP editor:

![image](https://github.com/user-attachments/assets/67a3c8d7-7ea5-4fa4-8384-debebc893233)

And in the CLspEditor it enables a quick switch back to the
traditional CEditor.

![image](https://github.com/user-attachments/assets/e508a726-f892-47da-92bd-a195bedc0dda)

If there is no project specific setting for `prefer_lsp` a
quick confirmation dialog is displayed with options for help
and feedback:

![image](https://github.com/user-attachments/assets/7c33d4fe-69a8-4cf8-b1fb-7ca80e07020c)

and the other way:

![image](https://github.com/user-attachments/assets/a69a87d8-4f32-4e58-a651-2583db50183e)

If the editor is open on a specific project and that project
has project specific preferences for `prefer_lsp`, instead of
automatically changing the setting, the property page is opened:

![image](https://github.com/user-attachments/assets/16c2ce3b-5b62-4199-8392-5d6800f621c7)

A notification is also displayed to the user for a short while after the editors are
re-opening offering more information and access to the preferences.

![image](https://github.com/user-attachments/assets/e8d069de-0939-4503-8755-a130595570d0)

A notification is also displayed when closing the banner allowing
the user easy access to disable opening the banner on new editors

![image](https://github.com/user-attachments/assets/9f9a00ef-8957-4a8a-8a3d-9c804eee5b5e)

Finally a new preference is available:

![image](https://github.com/user-attachments/assets/74d2d461-d219-465b-9276-e93f6720a9ed)

Fixes https://github.com/eclipse-cdt/cdt/issues/968