Improve triggering of license checks
License checks fail a lot due to rate limiting, and we often
run license checks on PRs that aren't changing dependencies,
which mean they *shouldn't* change license check.

Therefore only run license check if pom.xml/*.target files change
on a PR. For pushes, run the license check always so that we can
always see that the latest HEAD commit still is properly licensed.

Additionally, run on all push branches, but limit the runs to
the origin eclipse-cdt owned repos.

This sycnchronizes with the same changes made to cdt in
https://github.com/eclipse-cdt/cdt/pull/1388